adds fallback for convolveNN functions
