+++ src/backend/cuda/handle.cpp
@@ -9,8 +9,6 @@
 
 #include <common/unique_handle.hpp>
 #include <cublas.hpp>
-#include <cudnn.hpp>
-#include <cudnnModule.hpp>
 #include <cufft.hpp>
 #include <cusolverDn.hpp>
 #include <cusparse.hpp>
@@ -21,9 +19,17 @@ CREATE_HANDLE(cusparseHandle_t, cusparseCreate, cusparseDestroy);
 CREATE_HANDLE(cublasHandle_t, cublasCreate, cublasDestroy);
 CREATE_HANDLE(cusolverDnHandle_t, cusolverDnCreate, cusolverDnDestroy);
 CREATE_HANDLE(cufftHandle, cufftCreate, cufftDestroy);
+
+#ifdef WITH_CUDNN
+
+#include <cudnn.hpp>
+#include <cudnnModule.hpp>
+
 CREATE_HANDLE(cudnnHandle_t, cuda::getCudnnPlugin().cudnnCreate, cuda::getCudnnPlugin().cudnnDestroy);
 CREATE_HANDLE(cudnnTensorDescriptor_t, cuda::getCudnnPlugin().cudnnCreateTensorDescriptor, cuda::getCudnnPlugin().cudnnDestroyTensorDescriptor);
 CREATE_HANDLE(cudnnFilterDescriptor_t, cuda::getCudnnPlugin().cudnnCreateFilterDescriptor, cuda::getCudnnPlugin().cudnnDestroyFilterDescriptor);
 CREATE_HANDLE(cudnnConvolutionDescriptor_t, cuda::getCudnnPlugin().cudnnCreateConvolutionDescriptor, cuda::getCudnnPlugin().cudnnDestroyConvolutionDescriptor);
+
+#endif
 
 // clang-format on
