+++ src/backend/cuda/wrap.hpp
@@ -14,4 +14,10 @@ template<typename T>
 void wrap(Array<T> &out, const Array<T> &in, const dim_t ox, const dim_t oy,
           const dim_t wx, const dim_t wy, const dim_t sx, const dim_t sy,
           const dim_t px, const dim_t py, const bool is_column);
-}
+
+template<typename T>
+Array<T> wrap_dilated(const Array<T> &in, const dim_t ox, const dim_t oy,
+                      const dim_t wx, const dim_t wy, const dim_t sx,
+                      const dim_t sy, const dim_t px, const dim_t py,
+                      const dim_t dx, const dim_t dy, const bool is_column);
+}  // namespace cuda
